Trekk ut vedtakFellesFelter. Opprett brev util.
