--- src/test/kotlin/no/nav/familie/ba/sak/dokument/DokumentUtilsTest.kt
@@ -1,82 +1,0 @@
-package no.nav.familie.ba.sak.dokument
-
-import no.nav.familie.ba.sak.behandling.steg.StegType
-import no.nav.familie.ba.sak.common.lagBehandling
-import no.nav.familie.ba.sak.totrinnskontroll.domene.Totrinnskontroll
-import org.junit.Assert.assertEquals
-import org.junit.jupiter.api.Test
-
-internal class DokumentUtilsTest {
-
-    /**
-     * Siden navnet til saksbehandler blir hentet fra sikkerhetscontext er det riktig at denne er system her.
-     */
-    @Test
-    fun `Saksbehandler blir hentet fra sikkerhetscontext og beslutter viser placeholder tekst under behandling`() {
-        val behandling = lagBehandling()
-
-        val (saksbehandler, beslutter) = DokumentUtils.hentSaksbehandlerOgBeslutter(
-                behandling = behandling,
-                totrinnskontroll = null
-        )
-
-        assertEquals("System", saksbehandler)
-        assertEquals("Beslutter", beslutter)
-    }
-
-    @Test
-    fun `Saksbehandler blir hentet og beslutter er hentet fra sikkerhetscontext under beslutning`() {
-        val behandling = lagBehandling()
-        behandling.leggTilBehandlingStegTilstand(StegType.BESLUTTE_VEDTAK)
-
-        val (saksbehandler, beslutter) = DokumentUtils.hentSaksbehandlerOgBeslutter(
-                behandling = behandling,
-                totrinnskontroll = Totrinnskontroll(
-                        behandling = behandling,
-                        saksbehandler = "Mock Saksbehandler",
-                        saksbehandlerId = "mock.saksbehandler@nav.no"
-                )
-        )
-
-        assertEquals("Mock Saksbehandler", saksbehandler)
-        assertEquals("System", beslutter)
-    }
-
-    @Test
-    fun `Saksbehandler blir hentet og beslutter viser placeholder tekst under beslutning`() {
-        val behandling = lagBehandling()
-        behandling.leggTilBehandlingStegTilstand(StegType.BESLUTTE_VEDTAK)
-
-        val (saksbehandler, beslutter) = DokumentUtils.hentSaksbehandlerOgBeslutter(
-                behandling = behandling,
-                totrinnskontroll = Totrinnskontroll(
-                        behandling = behandling,
-                        saksbehandler = "System",
-                        saksbehandlerId = "systembruker"
-                )
-        )
-
-        assertEquals("System", saksbehandler)
-        assertEquals("Beslutter", beslutter)
-    }
-
-    @Test
-    fun `Saksbehandler og beslutter blir hentet etter at totrinnskontroll er besluttet`() {
-        val behandling = lagBehandling()
-        behandling.leggTilBehandlingStegTilstand(StegType.BESLUTTE_VEDTAK)
-
-        val (saksbehandler, beslutter) = DokumentUtils.hentSaksbehandlerOgBeslutter(
-                behandling = behandling,
-                totrinnskontroll = Totrinnskontroll(
-                        behandling = behandling,
-                        saksbehandler = "Mock Saksbehandler",
-                        saksbehandlerId = "mock.saksbehandler@nav.no",
-                        beslutter = "Mock Beslutter",
-                        beslutterId = "mock.beslutter@nav.no"
-                )
-        )
-
-        assertEquals("Mock Saksbehandler", saksbehandler)
-        assertEquals("Mock Beslutter", beslutter)
-    }
-}
